feat and fix
add methods to create, get, update and delete products in ProductService. In ProductRepository add find method and also in Product intitie on attribute name make this name unique in the db.

In Category modify the list of products with @JsonBackReference.

In UserController in the function getAllUsers in @PreAuthorize change to SUPER_ADMIN_ROLE

In ProductController add endpoints like getAllProduct, AddProduct, updateProductById, DeleteProductById

In AuhtRestController remove the sigup.